Optimisations (measured +400k nodes per second search speed on my laptop). Improve pawn move generation, change mobility eval to use simplified movegen for counting only, pack transposition table entries down to 16 bytes (from 24) (now 4 per cache line), switch all instances of '1ul' to '1ull' for improved portability

+++ lib/src/core.h
@@ -60,7 +60,7 @@ struct Piece {
             default: return "None";
         }
     }
-    static int64_t Evaluate(Value value) 
+    static int32_t Evaluate(Value value) 
     {
         switch (value) {
             case PAWN: return 100;
